requirements.txt
Pillow==10.1.0
qrcode==7.4.2
cryptography==41.0.7
playwright==1.40.0
